Shuffle around helper functions, eliminate redundancy.
